eLearning/requirements.txt
asgiref==3.5.1
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
certifi==2022.5.18.1
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
daphne==3.0.2
dj-rest-auth==2.2.4
Django==4.0.4
django-admin-volt==0.0.5
django-channels-presence==1.0.0
django-cors-headers==3.13.0
django-volt-admin==1.0.2
djangochannelsrestframework==1.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.1.0
MarkupSafe==2.1.1
mysqlclient==2.1.0
nltk==3.7
numpy==1.22.4
packaging==21.3
Pillow==9.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
regex==2022.6.2
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.64.0
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.2.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
websocket-client==1.3.2
zope.interface==5.4.0
